Added basic CAT support
The CAT was a co-ordinating assistant tracker in Kamaelia. In practice it
was fundamentally a key/value lookup store that was originally envisioned to
be local to components and sub components, but was mostly used and most useful
as a global lookup mechanism. As a result, I've added in a mechanism to register
values (which may be actors) with names that can be looked up across the system.

Since this Actor system is entirely thread based, the implementation here uses
the STM code developed for Kamaelia - for basic thread safety reasons.