One way to model a traditional accounts
This example is based on comments in "Sane concurrency in Go", and
rewriting it based on how the code should work. Worth noting that this
shows how to do threadsafe exceptions *between* threads.

Unlike the code / example here:

   https://blog.mozilla.org/services/2014/03/12/sane-concurrency-with-go/

I feel that this example is somewhat more readable, while being just as safe.