Process incoming messages before main loop
These were done after the main loop. However this led to to some minor
oddities - such as if an actor has a number of late bound methods, you
needed a matching number of yields before you could assume they've been
dealt with.

It will become clear at some point whether this is a good idea or a bad
one, but for now it's a better decision than the one before.