Allow use of main() as well as gen_process()
It's more intuitive to use main() as the main loop (if any) of the
actor. More intuitive than gen_process() at least - so you can
now use main() instead. At some point gen_process() will become
deprecated, but it's been the core name for a while, so it can stay for
now