Ensure that generator in main thread is a generator
In order to run the thread mainloop, we run a generator alongside the
interactions needed for the actor (outbound/inbound queues).

Given that the function the user supplies must be a generator or else
there will be subtle errors that start creeping in. So we now enforce
this by checking the code object before starting it. Raises a ValueError
(which is perhaps wrong error) if the function isn't a generator (which
is technically a value error, but perhaps something better might be
useful)