Evolve mini-guild example towards something better?
Tentative changes to see if we can simplify the minimal guild
reimplementation so that it can be used for teaching ideas around guild and
further experimentation.

Mini-Axon was a tutorial created for Kamaelia that was originally intended
to get a novice python developer up to speed with the techniques we used in
Kamaelia, but also was useful for testing ideas too.  This may have similar
benefits.

Changes:

* No longer an Actor and Activity base class - just Actor

* Activity has been simplified to be effectively an interface specification,
  called API, currently containing just input/output.  It or somthing like
  it is likely to be necessary in a more complete mini/micro guild
  implementation.

* Actor writers now only have to write a behaviour class.  A behaviour class
  is just a class with a bunch of methods, nothing special (yet).

* Behaviours can be stateless or stateful and have a thread of control or
  not.  Currently actor writers are responsible for ensuring their actors
  greenthread is started appropriately.  This is likely to change.

* Behaviours are enclosed by Actors, and can be written in a compact form,
  or separate from the Actor enclosure.