Update mini-guild to support multiple behaviors for an actor
Add support for multiple behaviours in an actor.  This is based on an akka
door open/closing example as a starting point.  Behaviours all require
access to the initialiser arguments - so they can be initialised
appropriately.

Also adds support for custom actor methods.  Initial test for this was to
add a munch method as well as input in the Consumer actor.