Modified version of accounts-3.py designed to demonstrate race hazards
This turns out to be harder than expected - which is why microsleeps
have been thrown in - to force opportunities for broken interleaving.
The races manifest as results where the amount of money in the system
at the end of the run is different from the amount of money at the start.