Re-enable actors with just a main() generator
This got bust, which is annoying. It's actually a logic error
overall, and ideally the fix would be more than this fix, but
this fix will work for the bulk of situations. What would be better
is to seperate out the 3 criterion:
* Is the process_method running
* Is there a generator and is it running
* Have we been asked to shutdown.

At the moment these 3 are conflated which is a problem, but this
change fixes this for the bulk of situations. Would be nice to
rewrite this part of the codebase.